- added spaces in css - removed console log - cleaned up tests - removed unneeded story - Updated togglebutton css - Removed extra line - Changed classes
