+++ packages/react-ui-core/src/Button/__tests__/mocks/theme.js
@@ -2,6 +2,6 @@ import { keyMirror } from '@rentpath/react-ui-utils'
 
 export default keyMirror([
   'ToggleButton',
-  'ToggleButton-On',
-  'ToggleButton-Off',
+  'ToggleButton-on',
+  'ToggleButton-off',
 ])
