+++ packages/react-ui-core/src/Button/index.js
@@ -1,2 +1,3 @@
 export { default } from './Button'
 export { default as Button } from './Button'
+export { default as ToggleButton } from './ToggleButton'
