--- packages/react-ui-core/src/ToggleButton/index.js
@@ -1,1 +1,0 @@
-export { default as ToggleButton } from './ToggleButton'
