+++ stories/theme/core/ToggleButton.css
@@ -1,16 +1,18 @@
 .ToggleButton {
   text-align: center;
-  color: darkgrey;
+  padding: 0;
+  border: 0 none;
+  background-color: transparent;
+  color: #fff;
+  box-shadow: 0 0 0 0 rgba(0, 0, 0, 0);
 }
 
-.ToggleButton-On svg{
+.ToggleButton-on svg {
   fill: #4A8FE2;
   stroke: #4A8FE2;
 }
 
-.ToggleButton-Off svg{
+.ToggleButton-off svg {
   fill: transparent;
   stroke: #4A8FE2;
 }
-
-
